+++ src/Provider/Provider.php
@@ -1,0 +1,12 @@
+<?php
+declare( strict_types=1 );
+
+namespace Rarst\wps\Provider;
+
+interface Provider {
+
+	public function get_name(): string;
+
+	public function get_data(): array;
+
+}