+++ wps.php
@@ -1,14 +1,16 @@
 <?php
 /*
 Plugin Name: wps
-Plugin URI: https://github.com/Rarst/wps
+Plugin URI: https://github.com/bart-jaskulski/wps
 Description: WordPress plugin for Whoops error handler.
-Author: Andrey "Rarst" Savchenko
-Version: 
+Author: Andrey "Rarst" Savchenko, Bartek Jaskulski
+Version: 2.0.0
 Author URI: http://www.rarst.net/
 License: MIT
+Requires at least: 5.9
 
 Copyright (c) 2013 Andrey "Rarst" Savchenko
+Copyright (c) 2023 Bartek Jaskulski
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this
 software and associated documentation files (the "Software"), to deal in the Software
@@ -27,13 +29,28 @@ OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 DEALINGS IN THE SOFTWARE.
 */
 
-if ( file_exists( __DIR__ . '/vendor/autoload.php' ) ) {
-	require __DIR__ . '/vendor/autoload.php';
+// Bail early on a series of preliminary checks.
+if ( ! defined( 'ABSPATH' ) ) {
+	return;
 }
 
+if ( ! defined( 'WP_DEBUG' ) || ! WP_DEBUG ) {
+	return;
+}
+
+if ( ! defined( 'WP_DEBUG_DISPLAY' ) || ! WP_DEBUG_DISPLAY ) {
+	return;
+}
+
+// Special GET parameter to disable wps.
 if ( isset( $_GET['wps_disable'] ) ) {
 	return;
 }
 
-$wps = new \Rarst\wps\Plugin();
+require __DIR__ . '/vendor/autoload.php';
+
+$container = new \Rarst\wps\Vendor\Pimple\Container();
+$container->register( new \Rarst\wps\ServiceProvider() );
+$wps = new \Rarst\wps\Plugin( $container[ \Rarst\wps\Vendor\Whoops\Run::class ] );
+do_action( 'wps/loaded', $wps );
 $wps->run();